$(cat <<'EOF'
