+++ StarcUp.UI/.gitignore
@@ -25,3 +25,4 @@ dist-ssr
 
 dist-electron
 
+public/resources
