+++ StarcUp.UI/src/main-page/vite-env.d.ts
@@ -1,0 +1,17 @@
+/// <reference types="vite/client" />
+
+declare global {
+  interface Window {
+    electronAPI: {
+      minimizeWindow: () => void
+      maximizeWindow: () => void
+      closeWindow: () => void
+      dragWindow: () => void
+      toggleOverlay: () => void
+      showOverlay: () => void
+      hideOverlay: () => void
+    }
+  }
+}
+
+export {}
