--- StarcUp.UI/src/vite-env.d.ts
@@ -1,15 +1,0 @@
-/// <reference types="vite/client" />
-
-declare global {
-  interface Window {
-    electronAPI: {
-      minimizeWindow: () => void
-      maximizeWindow: () => void
-      closeWindow: () => void
-      dragWindow: () => void
-      toggleOverlay: () => void
-      showOverlay: () => void
-      hideOverlay: () => void
-    }
-  }
-}
