+++ StarcUp.UI/vite.config.ts
@@ -3,12 +3,21 @@ import path from 'node:path'
 import electron from 'vite-plugin-electron/simple'
 import react from '@vitejs/plugin-react'
 import tailwindcss from '@tailwindcss/vite'
+import { build } from 'vite-plugin-electron'
 
 // https://vitejs.dev/config/
 export default defineConfig({
+  build: {
+    rollupOptions: {
+      input: {
+        main: path.resolve(__dirname, 'src/main-page/index.html'),
+        overlay: path.resolve(__dirname, 'src/overlay/index.html')
+      }
+    }
+  },
   plugins: [
     tailwindcss(),
-    react(),    
+    react(),
     electron({
       main: {
         // Shortcut of `build.lib.entry`.
